Ensure index values stored as int32

+++ improver/pollen/pollen_index_for_period.py
@@ -58,7 +58,7 @@ class PollenIndexForPeriod(PostProcessingPlugin):
         # Set values which are masked in _output_cube to nan
         self._output_cube.data = np.where(
             np.isnan(input_data), np.nan, self._output_cube.data
-        )
+        ).astype(np.int32)
 
     def _metadata(self, taxa: str):
         """Change the cube name and other metadata.
